removed try/block syntax, changed float to qreal

+++ VPVM/video/AudioDecoder.h
@@ -60,7 +60,7 @@ public:
 
 protected:
     virtual void run();
-    virtual void decodeBuffer(const QByteArray &bytes, float position, int channels);
+    virtual void decodeBuffer(const QByteArray &bytes, qreal position, int channels);
 
 signals:
     void audioDidDecode(const QByteArray &bytes);
